PyTorch/InsightFaceDetec/model/requirements.txt
tqdm
numpy
scipy
onnx 
onnxruntime 
onnx-simplifier
six
cython
scikit-image
pycocotools
torch==1.10.2
torchvision==0.11.3 
mmcv-full==1.3.3
terminaltables 